Adds the buttons and prepares for the integration
